Move two util functions from actuator to delete_in_batch, where they are more appropriate
